hw6 solved

+++ app/controllers/api/bookings_controller.rb
@@ -1,11 +1,20 @@
 module Api
   class BookingsController < ApplicationController
     def index
-      render json: jsonapi_serializer? ? jsonapi_booking(Booking.all) : blueprinter_all_bookings
+      booking = policy_scope(Booking)
+
+      render json: if jsonapi_serializer?
+                     jsonapi_booking(booking)
+                   else
+                     default_json_booking(booking)
+                   end
     end
 
     def show
       booking = Booking.find(params[:id])
+
+      authorize booking
+
       render json: if jsonapi_serializer?
                      jsonapi_booking(booking)
                    else
@@ -14,7 +23,10 @@ module Api
     end
 
     def create
-      booking = Booking.new(booking_params)
+      booking = Booking.new(merged_booking_params)
+
+      authorize booking
+
       if booking.save
         render json: default_json_booking(booking), status: :created
       else
@@ -24,6 +36,9 @@ module Api
 
     def update
       booking = Booking.find(params[:id])
+
+      authorize booking
+
       if booking.update(booking_params)
         render json: default_json_booking(booking), status: :ok
       else
@@ -33,6 +48,9 @@ module Api
 
     def destroy
       booking = Booking.find(params[:id])
+
+      authorize booking
+
       if booking.destroy
         render json: booking, status: :no_content
       else
@@ -43,27 +61,41 @@ module Api
     private
 
     def booking_params
-      params.require(:booking).permit(:id,
-                                      :flight_id,
+      if current_user&.admin?
+        admin_user_params
+      else
+        regular_user_params
+      end
+    end
+
+    def regular_user_params
+      params.require(:booking).permit(:flight_id,
+                                      :no_of_seats,
+                                      :seat_price)
+    end
+
+    def admin_user_params
+      params.require(:booking).permit(:flight_id,
                                       :no_of_seats,
                                       :seat_price,
                                       :user_id)
     end
 
     def default_json_booking(booking)
-      BookingSerializer.render(booking, root: :booking, view: :normal)
+      if root?
+        root_name = booking.is_a?(Booking) ? :booking : :bookings
+        BookingSerializer.render(booking, root: root_name, view: :normal)
+      else
+        BookingSerializer.render(booking, view: :normal)
+      end
+    end
+
+    def merged_booking_params
+      { 'user_id' => @current_user&.id }.merge(booking_params)
     end
 
     def jsonapi_booking(booking)
       JsonapiSerializer::BookingSerializer.new(booking).public_send(json_root_method)
-    end
-
-    def blueprinter_all_bookings
-      if root?
-        BookingSerializer.render(Booking.all, root: :bookings, view: :normal)
-      else
-        BookingSerializer.render(Booking.all, view: :normal)
-      end
     end
   end
 end
